+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## Unreleased
 
-* Changed
+* Added
   * Migrate doc from terra-card.
+
+* Added
+  * Migrate doc from terra-button.
+
+* Added
+  * Migrate doc from terra-button-group.
 
 ## 1.11.0 - (October 8, 2021)